implement MutCMapFromSet, add proper ResettingMutCMap
