+++ client/src/crdt/map/index.ts
@@ -3,5 +3,7 @@ export * from "./deleting_mut_map";
 export * from "./implicit_mut_map";
 export * from "./interfaces";
 export * from "./lww_map";
+export * from "./merging_mut_map";
+export * from "./mut_map_from_set";
 export * from "./resetting_mut_map";
 export * from "./tombstone_mut_map";
